backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
baato==1.0.1
certifi==2026.2.25
charset-normalizer==3.4.5
chromadb>=1.4.0
decorator==5.2.1
distro==1.9.0
Django==4.2.29
django-cors-headers==4.9.0
djangorestframework==3.16.1
#groq==0.9.0
gunicorn==20.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==0.19.2
requests==2.32.5
setuptools==82.0.1
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
#google-generativeai==0.3.0